+++ aldrin/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `introspection` Cargo feature to enable introspection support.
-- Add `RunError::Serialize` variant.
+- Add `RunError::Serialize` and `Deserialize` variants.
 - Support introspection.
 
 ### Changed